+++ dev/src/v1/firestore_admin_client.js
@@ -52,12 +52,22 @@ class FirestoreAdminClient {
    *     your project ID will be detected automatically.
    * @param {function} [options.promise] - Custom promise module to use instead
    *     of native Promises.
-   * @param {string} [options.servicePath] - The domain name of the
+   * @param {string} [options.apiEndpoint] - The domain name of the
    *     API remote host.
    */
   constructor(opts) {
     opts = opts || {};
     this._descriptors = {};
+
+    if (global.isBrowser) {
+      // If we're in browser, we use gRPC fallback.
+      opts.fallback = true;
+    }
+
+    // If we are in browser, we are already using fallback because of the
+    // "browser" field in package.json.
+    // But if we were explicitly requested to use fallback, let's do it now.
+    const gaxModule = !global.isBrowser && opts.fallback ? gax.fallback : gax;
 
     const servicePath =
       opts.servicePath || opts.apiEndpoint || this.constructor.servicePath;
@@ -75,42 +85,57 @@ class FirestoreAdminClient {
     // Create a `gaxGrpc` object, with any grpc-specific options
     // sent to the client.
     opts.scopes = this.constructor.scopes;
-    const gaxGrpc = new gax.GrpcClient(opts);
+    const gaxGrpc = new gaxModule.GrpcClient(opts);
 
     // Save the auth object to the client, for use by other methods.
     this.auth = gaxGrpc.auth;
 
     // Determine the client header string.
-    const clientHeader = [
-      `gl-node/${process.version}`,
-      `grpc/${gaxGrpc.grpcVersion}`,
-      `gax/${gax.version}`,
-      `gapic/${VERSION}`,
-    ];
+    const clientHeader = [];
+
+    if (typeof process !== 'undefined' && 'versions' in process) {
+      clientHeader.push(`gl-node/${process.versions.node}`);
+    }
+    clientHeader.push(`gax/${gaxModule.version}`);
+    if (opts.fallback) {
+      clientHeader.push(`gl-web/${gaxModule.version}`);
+    } else {
+      clientHeader.push(`grpc/${gaxGrpc.grpcVersion}`);
+    }
+    clientHeader.push(`gapic/${VERSION}`);
     if (opts.libName && opts.libVersion) {
       clientHeader.push(`${opts.libName}/${opts.libVersion}`);
     }
 
     // Load the applicable protos.
+    // For Node.js, pass the path to JSON proto file.
+    // For browsers, pass the JSON content.
+
+    const nodejsProtoPath = path.join(
+      __dirname,
+      '..',
+      '..',
+      'protos',
+      'protos.json'
+    );
     const protos = gaxGrpc.loadProto(
-      path.join(__dirname, '..', '..', 'protos'),
-      ['google/firestore/admin/v1/firestore_admin.proto']
+      opts.fallback ? require('../../protos/protos.json') : nodejsProtoPath
     );
 
     // This API contains "path templates"; forward-slash-separated
     // identifiers to uniquely identify resources within the API.
     // Create useful helper objects for these.
     this._pathTemplates = {
-      databasePathTemplate: new gax.PathTemplate(
+      databasePathTemplate: new gaxModule.PathTemplate(
         'projects/{project}/databases/{database}'
       ),
-      fieldPathTemplate: new gax.PathTemplate(
+      fieldPathTemplate: new gaxModule.PathTemplate(
         'projects/{project}/databases/{database}/collectionGroups/{collection_id}/fields/{field_id}'
       ),
-      indexPathTemplate: new gax.PathTemplate(
+      indexPathTemplate: new gaxModule.PathTemplate(
         'projects/{project}/databases/{database}/collectionGroups/{collection_id}/indexes/{index_id}'
       ),
-      parentPathTemplate: new gax.PathTemplate(
+      parentPathTemplate: new gaxModule.PathTemplate(
         'projects/{project}/databases/{database}/collectionGroups/{collection_id}'
       ),
     };
@@ -119,12 +144,12 @@ class FirestoreAdminClient {
     // (e.g. 50 results at a time, with tokens to get subsequent
     // pages). Denote the keys used for pagination and results.
     this._descriptors.page = {
-      listIndexes: new gax.PageDescriptor(
+      listIndexes: new gaxModule.PageDescriptor(
         'pageToken',
         'nextPageToken',
         'indexes'
       ),
-      listFields: new gax.PageDescriptor(
+      listFields: new gaxModule.PageDescriptor(
         'pageToken',
         'nextPageToken',
         'fields'
@@ -147,7 +172,9 @@ class FirestoreAdminClient {
     // Put together the "service stub" for
     // google.firestore.admin.v1.FirestoreAdmin.
     const firestoreAdminStub = gaxGrpc.createStub(
-      protos.google.firestore.admin.v1.FirestoreAdmin,
+      opts.fallback
+        ? protos.lookupService('google.firestore.admin.v1.FirestoreAdmin')
+        : protos.google.firestore.admin.v1.FirestoreAdmin,
       opts
     );
 
@@ -165,18 +192,16 @@ class FirestoreAdminClient {
       'updateField',
     ];
     for (const methodName of firestoreAdminStubMethods) {
-      this._innerApiCalls[methodName] = gax.createApiCall(
-        firestoreAdminStub.then(
-          stub =>
-            function() {
-              const args = Array.prototype.slice.call(arguments, 0);
-              return stub[methodName].apply(stub, args);
-            },
-          err =>
-            function() {
-              throw err;
-            }
-        ),
+      const innerCallPromise = firestoreAdminStub.then(
+        stub => (...args) => {
+          return stub[methodName].apply(stub, args);
+        },
+        err => () => {
+          throw err;
+        }
+      );
+      this._innerApiCalls[methodName] = gaxModule.createApiCall(
+        innerCallPromise,
         defaults[methodName],
         this._descriptors.page[methodName]
       );
@@ -282,6 +307,7 @@ class FirestoreAdminClient {
       callback = options;
       options = {};
     }
+    request = request || {};
     options = options || {};
     options.otherArgs = options.otherArgs || {};
     options.otherArgs.headers = options.otherArgs.headers || {};
@@ -386,6 +412,7 @@ class FirestoreAdminClient {
       callback = options;
       options = {};
     }
+    request = request || {};
     options = options || {};
     options.otherArgs = options.otherArgs || {};
     options.otherArgs.headers = options.otherArgs.headers || {};
@@ -498,6 +525,7 @@ class FirestoreAdminClient {
       callback = options;
       options = {};
     }
+    request = request || {};
     options = options || {};
     options.otherArgs = options.otherArgs || {};
     options.otherArgs.headers = options.otherArgs.headers || {};
@@ -544,6 +572,7 @@ class FirestoreAdminClient {
       callback = options;
       options = {};
     }
+    request = request || {};
     options = options || {};
     options.otherArgs = options.otherArgs || {};
     options.otherArgs.headers = options.otherArgs.headers || {};
@@ -611,6 +640,7 @@ class FirestoreAdminClient {
       callback = options;
       options = {};
     }
+    request = request || {};
     options = options || {};
     options.otherArgs = options.otherArgs || {};
     options.otherArgs.headers = options.otherArgs.headers || {};
@@ -683,6 +713,7 @@ class FirestoreAdminClient {
       callback = options;
       options = {};
     }
+    request = request || {};
     options = options || {};
     options.otherArgs = options.otherArgs || {};
     options.otherArgs.headers = options.otherArgs.headers || {};
@@ -737,6 +768,7 @@ class FirestoreAdminClient {
       callback = options;
       options = {};
     }
+    request = request || {};
     options = options || {};
     options.otherArgs = options.otherArgs || {};
     options.otherArgs.headers = options.otherArgs.headers || {};
@@ -850,6 +882,7 @@ class FirestoreAdminClient {
       callback = options;
       options = {};
     }
+    request = request || {};
     options = options || {};
     options.otherArgs = options.otherArgs || {};
     options.otherArgs.headers = options.otherArgs.headers || {};
@@ -984,6 +1017,7 @@ class FirestoreAdminClient {
       callback = options;
       options = {};
     }
+    request = request || {};
     options = options || {};
     options.otherArgs = options.otherArgs || {};
     options.otherArgs.headers = options.otherArgs.headers || {};