+++ dev/src/v1beta1/doc/google/firestore/v1beta1/doc_query.js
@@ -239,26 +239,19 @@ const StructuredQuery = {
        * Contains. Requires that the field is an array.
        */
       ARRAY_CONTAINS: 7,
-    },
-  },
 
-  /**
-   * The projection of document's fields to return.
-   *
-   * @property {Object[]} fields
-   *   The fields to return.
-   *
-   *   If empty, all fields are returned. To only return the name
-   *   of the document, use `['__name__']`.
-   *
-   *   This object should have the same structure as [FieldReference]{@link google.firestore.v1beta1.FieldReference}
-   *
-   * @typedef Projection
-   * @memberof google.firestore.v1beta1
-   * @see [google.firestore.v1beta1.StructuredQuery.Projection definition in proto format]{@link https://github.com/googleapis/googleapis/blob/master/google/firestore/v1beta1/query.proto}
-   */
-  Projection: {
-    // This is for documentation. Actual contents will be loaded by gRPC.
+      /**
+       * In. Requires that `value` is a non-empty ArrayValue with at most 10
+       * values.
+       */
+      IN: 8,
+
+      /**
+       * Contains any. Requires that the field is an array and
+       * `value` is a non-empty ArrayValue with at most 10 values.
+       */
+      ARRAY_CONTAINS_ANY: 9,
+    },
   },
 
   /**
@@ -299,7 +292,7 @@ const StructuredQuery = {
       IS_NAN: 2,
 
       /**
-       * Test if an exprestion evaluates to Null.
+       * Test if an expression evaluates to Null.
        */
       IS_NULL: 3,
     },
@@ -323,6 +316,25 @@ const StructuredQuery = {
    * @see [google.firestore.v1beta1.StructuredQuery.Order definition in proto format]{@link https://github.com/googleapis/googleapis/blob/master/google/firestore/v1beta1/query.proto}
    */
   Order: {
+    // This is for documentation. Actual contents will be loaded by gRPC.
+  },
+
+  /**
+   * The projection of document's fields to return.
+   *
+   * @property {Object[]} fields
+   *   The fields to return.
+   *
+   *   If empty, all fields are returned. To only return the name
+   *   of the document, use `['__name__']`.
+   *
+   *   This object should have the same structure as [FieldReference]{@link google.firestore.v1beta1.FieldReference}
+   *
+   * @typedef Projection
+   * @memberof google.firestore.v1beta1
+   * @see [google.firestore.v1beta1.StructuredQuery.Projection definition in proto format]{@link https://github.com/googleapis/googleapis/blob/master/google/firestore/v1beta1/query.proto}
+   */
+  Projection: {
     // This is for documentation. Actual contents will be loaded by gRPC.
   },