+++ dev/test/gapic-v1beta1.js
@@ -15,7 +15,7 @@
 'use strict';
 
 const assert = require('assert');
-const through2 = require('through2');
+const {PassThrough} = require('stream');
 
 const firestoreModule = require('../src');
 
@@ -42,6 +42,13 @@ describe('FirestoreClient', () => {
 
   it('should create a client with no options', () => {
     const client = new firestoreModule.v1beta1.FirestoreClient();
+    assert(client);
+  });
+
+  it('should create a client with gRPC fallback', () => {
+    const client = new firestoreModule.v1beta1.FirestoreClient({
+      fallback: true,
+    });
     assert(client);
   });
 
@@ -1002,12 +1009,15 @@ function mockSimpleGrpcMethod(expectedRequest, response, error) {
 function mockServerStreamingGrpcMethod(expectedRequest, response, error) {
   return actualRequest => {
     assert.deepStrictEqual(actualRequest, expectedRequest);
-    const mockStream = through2.obj((chunk, enc, callback) => {
-      if (error) {
-        callback(error);
-      } else {
-        callback(null, response);
-      }
+    const mockStream = new PassThrough({
+      objectMode: true,
+      transform: (chunk, enc, callback) => {
+        if (error) {
+          callback(error);
+        } else {
+          callback(null, response);
+        }
+      },
     });
     return mockStream;
   };
@@ -1015,13 +1025,16 @@ function mockServerStreamingGrpcMethod(expectedRequest, response, error) {
 
 function mockBidiStreamingGrpcMethod(expectedRequest, response, error) {
   return () => {
-    const mockStream = through2.obj((chunk, enc, callback) => {
-      assert.deepStrictEqual(chunk, expectedRequest);
-      if (error) {
-        callback(error);
-      } else {
-        callback(null, response);
-      }
+    const mockStream = new PassThrough({
+      objectMode: true,
+      transform: (chunk, enc, callback) => {
+        assert.deepStrictEqual(chunk, expectedRequest);
+        if (error) {
+          callback(error);
+        } else {
+          callback(null, response);
+        }
+      },
     });
     return mockStream;
   };